refactor lecture request face id package
