+++ app/src/main/java/vn/edu/fpt/zentryapp/auth/ui/FaceIdSuccessActivity.java
@@ -1,8 +1,7 @@
-package com.example.zentry_app.ui.faceid;
+package vn.edu.fpt.zentryapp.auth.ui;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
